+++ CHANGELOG.md
@@ -2,12 +2,20 @@
 
 ## Release 0.0.15 (Unreleased)
 
+### Changed
+
+- Installation of dependencies from manifest is now optional.
+  It is frequent case, when authors of module place incorrect dependencies
+  in manifest, thus installation of addon may fail.
+
 ### Fixes
 
 - Fix error when running `addons install/update/uninstall` with non-existing
   logfile. This was caused by attempt to determine starting point of logfile
   to search for errors happened during operation.
   Now this case is handled correctly.
+
+---
 
 ## Release 0.0.14 (2023-10-04)